Report open networks without saved networks
The supplicant stops periodic scans when there are no saved networks.
The framework needs to have a periodic scan to handle this scenario.

We do have an infrequent wake up scan (15 mins), but thats way too
slow to report an open network.

Setup a scan by the supplicant interval when there are no saved
networks

Bug: 5420656
Change-Id: Id3708ecc874b42971643cc747bb9e7f2efc7d1dd