Sped up ActionBar and StatusBar animations
Animations to show/hide the ActionBar and StatusBar were very slow, given
the size of the objects and the distances covered by the sliding animations.
Also, the ActionBar animation was sometimes hiccuppy as it faded in/out.
This change eliminates the ActionBar fade (which is unnecessary) and speeds
up the animations (smaller durations and steeper interpolation curves). Also,
it eliminates the startDelay on the ActionBar show animation.

Issue #6564089 Options menu should slide in much quicker (nakasi/JB)

Change-Id: I2c8298301f7bf26bbbc94444e715420a2c029ba0