Make animations persist after completion.
Transition animations were set to fillAfter=false which caused them
to revert when they had completed. Normally this would not have
mattered as the final state was obscured by the new Activity. In
the case of the Phone app the Activities churn so much that the
original Activity was exposed.

Fixes bug 6691421.

Change-Id: I96b8a43d0a2cac9a34a14a309548c41451c65a57