Framwork: Add support for variable size pattern lockscreen (Part 1/2)
Changes LockPatternView to use PATTERN_SIZE instead of the hardcoded
value '3'.
Adds methods to set and get PATTERN_SIZE via LockPatternUtils and
device settings.

With this patch the pattern lockscreen can go from 3x3 to any sane size
such as 6x6.
LockPatternView can be taken by itself if settings are not desired, for
an easy way to change the hardcoded size.