SystemUI: Move statusbar container object
The original ICS thumbing code had the status bar container (to rebuild
the views when switching themes) in the SystemUI service itself, since
it was the only common point between the phone and tablet bars.

JB introduced a BaseStatusBar class from which both layout types
derive, so move the container (and its initialization) there.

Additionally, explicitly make the phonestatus barr view grab the input
focus when the window is expanded, to avoid missing the BACkey eventnt
that makes it collapse.

Change-Id: I288aa10ff0130f38d405aa8f81c124e0c06e7822