Lockscreen Vibrate option 1 of 2
Lockscreen Vibrate option 1 of 2

The new System setting Settings.System.LOCKSCREEN_VIBRATE_ENABLED
which defaults to 1, will effect the lock screen on touch event
(when touching the unlock icon - or sliding it).

The GlowPadView (Slider) now has no Vibrate when the user selects
thebehaviorr
in the Lockscreen Settings (see 2 of 2 patch).